refactor: extract `DbAdapter` to standalone class so it can refresh data correctly
refs #1306
